@@ -27,10 +13,96 @@ members = [
   "bomboni_wasm",
   "bomboni_wasm_core",
   "bomboni_wasm_derive",
+  "examples/grpc/grpc-common",
+  "examples/grpc/bookstore/bookstore-api",
+  "examples/grpc/bookstore/bookstore-service",
+  "examples/grpc/bookstore/bookstore-cli",
 ]
+exclude = ["integrations"]
+
+[workspace.package]
+version = "0.2.1"
+authors = ["Tin Rabzelj <tin@flinect.com>"]
+description = "Utility Library for Rust"
+repository = "https://github.com/tinrab/bomboni"
+homepage = "https://github.com/tinrab/bomboni"
+license-file = "LICENSE"
+edition = "2024"
+
+[workspace.dependencies]
+bomboni_common = { path = "bomboni_common", version = "0.2.1" }
+bomboni_core = { path = "bomboni_core", version = "0.2.1" }
+bomboni_fs = { path = "bomboni_fs", version = "0.2.1" }
+bomboni_macros = { path = "bomboni_macros", version = "0.2.1" }
+bomboni_prost = { path = "bomboni_prost", version = "0.2.1" }
+bomboni_proto = { path = "bomboni_proto", version = "0.2.1" }
+bomboni_request = { path = "bomboni_request", version = "0.2.1" }
+bomboni_request_derive = { path = "bomboni_request_derive", version = "0.2.1" }
+bomboni_template = { path = "bomboni_template", version = "0.2.1" }
+bomboni_wasm = { path = "bomboni_wasm", version = "0.2.1" }
+bomboni_wasm_core = { path = "bomboni_wasm_core", version = "0.2.1" }
+bomboni_wasm_derive = { path = "bomboni_wasm_derive", version = "0.2.1" }
+
+thiserror = "2.0.18"
+regex = "1.12.3"
+time = "0.3.47"
+chrono = "0.4.43"
+ulid = "1.2.1"
+bytes = "1.11.1"
+serde = "1.0.228"
+serde_json = "1.0.149"
+serde_derive_internals = "0.29.1"
+convert_case = "0.9.0"
+handlebars = "6.4.0"
+rand = "0.9.2"
+rand_core = "0.9.5"
+itertools = "0.14.0"
+paste = "1.0.15"
+
+http = "1.4.0"
+hyper = "1.8.1"
+prost = "0.14.3"
+prost-types = "0.14.3"
+prost-build = "0.14.3"
+tonic = "0.14.4"
+tonic-prost = "0.14.4"
+tonic-prost-build = "0.14.4"
+tonic-reflection = "0.14.4"
+tokio = "1.49.0"
+async-trait = "0.1.89"
+postgres-types = "0.2.12"
+mysql_common = "0.35.5"
+jsonwebtoken = "9.3"
+cookie = "0.18"
+
+tracing = "0.1.44"
+tracing-subscriber = "0.3.22"
+opentelemetry = "0.31.0"
+opentelemetry_sdk = "0.31.0"
+opentelemetry-stdout = "0.31.0"
+
+proc-macro2 = "1.0.106"
+syn = "2.0.116"
+quote = "1.0.44"
+prettyplease = "0.2.37"
+darling = "0.21.3"
+
+[package]
+name = "bomboni"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+repository.workspace = true
+homepage.workspace = true
+license-file.workspace = true
+readme = "README.md"
+edition.workspace = true
+
+[lib]
+name = "bomboni"
+path = "src/lib.rs"
 
 [features]
-default = ["prost", "proto", "request"]
 derive = ["bomboni_request/derive", "bomboni_wasm/derive"]
 prost = ["dep:bomboni_prost"]
 proto = ["prost", "dep:bomboni_proto"]
