@@ -1,3 +1,6 @@
+#![doc = include_str!("../README.md")]
+
+/// Date and time utilities.
 pub mod date_time;
+/// Unique identifier utilities.
 pub mod id;
-pub mod macros;
