@@ -1,5 +1,6 @@
 use syn::{Type, TypePath};
 
+/// Macro for formatting documentation comments.
 #[macro_export]
 macro_rules! format_comment {
     ($($arg:tt)*) => {{
@@ -10,21 +11,25 @@ macro_rules! format_comment {
     }};
 }
 
+/// Checks if a type is `PhantomData`.
+#[must_use]
 pub fn type_is_phantom(ty: &Type) -> bool {
     if let Type::Path(TypePath { path, .. }) = ty {
         path.segments
             .last()
-            .map_or(false, |path| path.ident == "PhantomData")
+            .is_some_and(|path| path.ident == "PhantomData")
     } else {
         false
     }
 }
 
+/// Checks if a type is `Option`.
+#[must_use]
 pub fn type_is_option(ty: &Type) -> bool {
     if let Type::Path(TypePath { path, .. }) = ty {
         path.segments
             .last()
-            .map_or(false, |path| path.ident == "Option")
+            .is_some_and(|path| path.ident == "Option")
     } else {
         false
     }
