@@ -1,13 +1,47 @@
+//! Oneof field code generation utilities.
+
 use std::collections::BTreeMap;
 
-use bomboni_core::string::{str_to_case, Case};
+use bomboni_core::string::{Case, str_to_case};
 use proc_macro2::{Ident, TokenStream};
-use prost_types::{field_descriptor_proto, DescriptorProto, OneofDescriptorProto};
+use prost_types::{DescriptorProto, OneofDescriptorProto, field_descriptor_proto};
 use quote::{format_ident, quote};
 use syn::TypePath;
 
-use crate::context::Context;
+use crate::{context::Context, utility::is_rust_keyword};
 
+/// Writes utility functions for oneof fields in a message.
+///
+/// This function generates helper functions for working with protobuf oneof fields,
+/// including getters, setters, and type checking utilities. These functions
+/// make it easier to work with oneof fields in Rust code.
+///
+/// # Arguments
+///
+/// * `context` - The generation context containing configuration
+/// * `s` - The token stream to write generated code to
+/// * `message` - The protobuf message descriptor containing oneof fields
+///
+/// # Generated Functions
+///
+/// For each oneof field, this generates:
+/// - Getter functions for each variant
+/// - Type checking functions
+/// - Utility functions for working with the oneof
+///
+/// # Examples
+///
+/// Given a protobuf message with:
+/// ```proto
+/// message User {
+///   oneof data {
+///     string email = 1;
+///     string phone = 2;
+///   }
+/// }
+/// ```
+///
+/// This function generates utility functions for working with the `data` oneof field.
 pub fn write_message_oneofs(context: &Context, s: &mut TokenStream, message: &DescriptorProto) {
     if message.oneof_decl.is_empty() {
         return;
@@ -53,10 +87,8 @@ fn write_name(
     oneof: &OneofDescriptorProto,
 ) {
     let message_ident = context.get_type_expr_path(message.name());
-    let oneof_name_ident = format_ident!(
-        "{}_ONEOF_NAME",
-        str_to_case(oneof.name(), Case::ScreamingSnake)
-    );
+    let oneof_name_ident =
+        format_ident!("{}_ONEOF_NAME", str_to_case(oneof.name(), Case::Constant));
     let oneof_name_literal = oneof.name();
     s.extend(quote! {
         impl #message_ident {
@@ -78,10 +110,8 @@ fn write_variant_names(
         .iter()
         .filter(|field| field.oneof_index == Some(oneof_index as i32))
     {
-        let variant_name_ident = format_ident!(
-            "{}_VARIANT_NAME",
-            str_to_case(field.name(), Case::ScreamingSnake)
-        );
+        let variant_name_ident =
+            format_ident!("{}_VARIANT_NAME", str_to_case(field.name(), Case::Constant));
         let variant_name_literal = field.name();
         variant_names.extend(quote! {
             pub const #variant_name_ident: &'static str = #variant_name_literal;
@@ -150,8 +180,10 @@ fn write_variant_from(
         }
         .to_string();
 
-        let variant_ident = format_ident!("{}", str_to_case(field.name(), Case::Pascal));
-        from_map.entry(source_type).or_default().push(variant_ident);
+        if field.type_name() != ".google.protobuf.Empty" {
+            let variant_ident = format_ident!("{}", str_to_case(field.name(), Case::Pascal));
+            from_map.entry(source_type).or_default().push(variant_ident);
+        }
     }
 
     // Only implement from if there is a single variant for the given type.
@@ -180,12 +212,18 @@ fn write_variant_from(
         {
             let message_ident = context.get_type_expr_path(message.name());
             let variant_ident = format_ident!("{}", str_to_case(oneof.name(), Case::Snake));
+            let variant_ident_str = variant_ident.to_string();
+            let escaped_variant_ident = if is_rust_keyword(&variant_ident_str) {
+                format_ident!("r#{}", variant_ident_str)
+            } else {
+                variant_ident
+            };
             s.extend(quote! {
                 /// From source variant type to owner message type.
                 impl From<#source_type> for #message_ident {
                     fn from(value: #source_type) -> Self {
                         Self {
-                            #variant_ident: Some(value.into()),
+                            #escaped_variant_ident: Some(value.into()),
                         }
                     }
                 }
@@ -207,10 +245,8 @@ fn write_variant_utility(
         .iter()
         .filter(|field| field.oneof_index == Some(oneof_index as i32))
     {
-        let variant_name_ident = format_ident!(
-            "{}_VARIANT_NAME",
-            str_to_case(field.name(), Case::ScreamingSnake)
-        );
+        let variant_name_ident =
+            format_ident!("{}_VARIANT_NAME", str_to_case(field.name(), Case::Constant));
         let oneof_field_ident = format_ident!("{}", str_to_case(field.name(), Case::Pascal));
         variant_cases.extend(quote! {
             Self::#oneof_field_ident(_) => Self::#variant_name_ident,
@@ -244,12 +280,18 @@ fn write_into_owner(context: &Context, s: &mut TokenStream, message: &Descriptor
     let oneof = message.oneof_decl.first().unwrap();
     let oneof_ident = context.get_oneof_ident(message, oneof);
     let variant_ident = format_ident!("{}", str_to_case(oneof.name(), Case::Snake));
+    let variant_ident_str = variant_ident.to_string();
+    let escaped_variant_ident = if is_rust_keyword(&variant_ident_str) {
+        format_ident!("r#{}", variant_ident_str)
+    } else {
+        variant_ident
+    };
 
     s.extend(quote! {
         impl From<#oneof_ident> for #message_ident {
             fn from(value: #oneof_ident) -> Self {
                 Self {
-                    #variant_ident: Some(value),
+                    #escaped_variant_ident: Some(value),
                 }
             }
         }
