@@ -7,8 +7,9 @@ use crate::google::rpc::{
 use crate::impl_proto_any_convert;
 
 impl Status {
+    /// Creates a new `Status` with the given code, message, and details.
     #[must_use]
-    pub fn new(code: Code, message: String, details: Vec<Any>) -> Self {
+    pub const fn new(code: Code, message: String, details: Vec<Any>) -> Self {
         Self {
             code: code as i32,
             message,
@@ -18,27 +19,31 @@ impl Status {
 }
 
 #[cfg(feature = "tonic")]
-impl From<tonic::Status> for Status {
-    fn from(status: tonic::Status) -> Self {
+impl TryFrom<tonic::Status> for Status {
+    type Error = prost::DecodeError;
+
+    fn try_from(status: tonic::Status) -> Result<Self, Self::Error> {
         use prost::Message;
-        let details = Self::decode(status.details()).unwrap();
-        Self {
+        let details = Self::decode(status.details())?;
+        Ok(Self {
             code: status.code() as i32,
             message: status.message().into(),
             details: details.details,
-        }
+        })
     }
 }
 
 #[cfg(feature = "tonic")]
-impl From<Status> for tonic::Status {
-    fn from(status: Status) -> Self {
+impl TryFrom<Status> for tonic::Status {
+    type Error = ();
+
+    fn try_from(status: Status) -> Result<Self, Self::Error> {
         use prost::Message;
         let code = tonic::Code::from(status.code);
         let message = status.message.clone();
         let mut encoded_details = Vec::new();
-        status.encode(&mut encoded_details).unwrap();
-        Self::with_details(code, message, encoded_details.into())
+        status.encode(&mut encoded_details).map_err(|_| ())?;
+        Ok(Self::with_details(code, message, encoded_details.into()))
     }
 }
 
@@ -96,12 +101,14 @@ mod tests {
         let s = Status::new(
             Code::InvalidArgument,
             "error".to_string(),
-            vec![Any::pack_from(&ErrorInfo {
-                reason: "a".to_string(),
-                domain: "b".to_string(),
-                metadata: BTreeMap::default(),
-            })
-            .unwrap()],
+            vec![
+                Any::from_msg(&ErrorInfo {
+                    reason: "a".to_string(),
+                    domain: "b".to_string(),
+                    metadata: BTreeMap::default(),
+                })
+                .unwrap(),
+            ],
         );
         let js = serde_json::to_string(&s).unwrap();
         assert_eq!(
