@@ -4,15 +4,15 @@ use crate::{
     query::{
         error::{QueryError, QueryResult},
         page_token::{
-            utility::{get_page_filter, make_page_key},
             FilterPageToken, PageTokenBuilder,
+            utility::{get_page_filter, make_page_key},
         },
     },
     schema::SchemaMapped,
 };
 use aes_gcm::{
-    aead::{Aead, OsRng},
     AeadCore, Aes256Gcm, Key, KeyInit,
+    aead::{Aead, OsRng},
 };
 use base64ct::{Base64, Base64Url, Encoding};
 use std::fmt::{self, Debug, Formatter};
@@ -28,7 +28,8 @@ pub struct Aes256PageTokenBuilder {
 }
 
 impl Aes256PageTokenBuilder {
-    pub fn new(url_safe: bool) -> Self {
+    /// Creates a new AES256 page token builder.
+    pub const fn new(url_safe: bool) -> Self {
         Self { url_safe }
     }
 }
