@@ -1,4 +1,5 @@
 //! The page token is used to determine the next page of results.
+//!
 //! How it is used to query the database is implementation-specific.
 //! One way is to filter IDs greater than the last item's ID of the previous page.
 //! If the query parameters change, then the page token is invalid.
@@ -10,29 +11,43 @@ use std::fmt::{self, Display, Formatter};
 use crate::{filter::Filter, ordering::Ordering, schema::SchemaMapped};
 
 use super::error::QueryResult;
+/// AES256 page token encoding.
 pub mod aes256;
+
+/// Base64 page token encoding.
 pub mod base64;
+
+/// Plain page token encoding.
 pub mod plain;
+
+/// RSA page token encoding.
 pub mod rsa;
 mod utility;
 
 /// A page token containing a filter.
 #[derive(Debug, Clone, PartialEq)]
 pub struct FilterPageToken {
+    /// Filter.
     pub filter: Filter,
 }
 
 impl FilterPageToken {
-    pub fn new(filter: Filter) -> Self {
+    /// Creates a new filter page token.
+    pub const fn new(filter: Filter) -> Self {
         Self { filter }
     }
 }
 
+/// Trait for building and parsing page tokens.
 pub trait PageTokenBuilder {
+    /// Page token type.
     type PageToken: Clone + ToString;
 
     /// Parse a page token.
-    /// [`QueryError::InvalidPageToken`] is returned if the page token is invalid for any reason.
+    ///
+    /// # Errors
+    ///
+    /// Will return an error if the page token cannot be parsed or is invalid.
     fn parse(
         &self,
         filter: &Filter,
@@ -45,7 +60,11 @@ pub trait PageTokenBuilder {
     ///
     /// Note that "last item" is not necessarily the last item of the page, but N+1th one.
     /// We can fetch `page_size+1` items from the database to determine if there are more results.
-    /// [`QueryError::PageTokenFailure`] is returned if the page token could not be built.
+    /// [`crate::query::error::QueryError::PageTokenFailure`] is returned if the page token could not be built.
+    ///
+    /// # Errors
+    ///
+    /// Will return an error if the page token cannot be built.
     fn build_next<T: SchemaMapped>(
         &self,
         filter: &Filter,
