@@ -1,16 +1,17 @@
 use crate::{
     filter::{
-        error::{FilterError, FilterResult},
         Filter, FilterComparator,
+        error::{FilterError, FilterResult},
     },
     schema::{FunctionSchemaMap, Schema, ValueType},
     sql::{
-        utility::{get_argument_parameter, get_identifier},
         SqlArgumentStyle, SqlDialect, SqlRenameMap,
+        utility::{get_argument_parameter, get_identifier},
     },
     value::Value,
 };
 
+/// Builder for SQL filter statements.
 pub struct SqlFilterBuilder<'a> {
     dialect: SqlDialect,
     argument_style: SqlArgumentStyle,
@@ -24,6 +25,7 @@ pub struct SqlFilterBuilder<'a> {
 }
 
 impl<'a> SqlFilterBuilder<'a> {
+    /// Creates a new SQL filter builder.
     pub fn new(dialect: SqlDialect, schema: &'a Schema) -> Self {
         Self {
             dialect,
@@ -38,31 +40,48 @@ impl<'a> SqlFilterBuilder<'a> {
         }
     }
 
-    pub fn set_schema_functions(&mut self, schema_functions: &'a FunctionSchemaMap) -> &mut Self {
+    /// Sets the schema functions.
+    pub const fn set_schema_functions(
+        &mut self,
+        schema_functions: &'a FunctionSchemaMap,
+    ) -> &mut Self {
         self.schema_functions = Some(schema_functions);
         self
     }
 
-    pub fn set_rename_map(&mut self, rename_map: &'a SqlRenameMap) -> &mut Self {
+    /// Sets the rename map.
+    pub const fn set_rename_map(&mut self, rename_map: &'a SqlRenameMap) -> &mut Self {
         self.rename_map = Some(rename_map);
         self
     }
 
-    pub fn set_document_offset(&mut self, offset: usize) -> &mut Self {
+    /// Sets the document offset.
+    pub const fn set_document_offset(&mut self, offset: usize) -> &mut Self {
         self.argument_offset = offset;
         self
     }
 
-    pub fn case_insensitive_like(&mut self) -> &mut Self {
+    /// Enables case insensitive like.
+    pub const fn case_insensitive_like(&mut self) -> &mut Self {
         self.case_insensitive_like = true;
         self
     }
 
+    /// Sets the argument style.
     pub fn set_argument_style(&mut self, argument_style: SqlArgumentStyle) -> &mut Self {
         self.argument_style = argument_style;
         self
     }
 
+    /// Builds a SQL filter.
+    ///
+    /// # Errors
+    ///
+    /// Will return [`FilterError::InvalidType`] if filter parts have incompatible types.
+    /// Will return [`FilterError::UnknownFunction`] if the filter contains an unknown function name.
+    /// Will return [`FilterError::UnknownMember`] if the filter contains an unknown field name.
+    /// Will return [`FilterError::IncomparableType`] if trying to compare incomparable types.
+    /// Will return [`FilterError::UnsuitableComparator`] if an unsuitable comparator is used.
     pub fn build(&mut self, filter: &Filter) -> FilterResult<(String, Vec<Value>)> {
         self.build_tree(filter)?;
 
@@ -383,7 +402,7 @@ mod tests {
     use std::collections::BTreeMap;
 
     use crate::{schema::FunctionSchema, testing::schema::RequestItem};
-    use bomboni_common::btree_map_into;
+    use bomboni_macros::btree_map_into;
 
     use super::*;
 
@@ -433,17 +452,25 @@ mod tests {
         assert_eq!(sql, r#"REGEX("user"."displayName", $1)"#);
         assert_eq!(args[0], Value::String("a".into()));
 
-        assert!(SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
-            .build(&Filter::parse("logs").unwrap())
-            .is_err());
-        assert!(SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
-            .build(&Filter::parse("user.logs").unwrap())
-            .is_err());
-        assert!(SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
-            .build(&Filter::parse("user.id = 42").unwrap())
-            .is_err());
-        assert!(SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
-            .build(&Filter::parse("task.deleted < false").unwrap())
-            .is_err());
+        assert!(
+            SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
+                .build(&Filter::parse("logs").unwrap())
+                .is_err()
+        );
+        assert!(
+            SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
+                .build(&Filter::parse("user.logs").unwrap())
+                .is_err()
+        );
+        assert!(
+            SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
+                .build(&Filter::parse("user.id = 42").unwrap())
+                .is_err()
+        );
+        assert!(
+            SqlFilterBuilder::new(SqlDialect::Postgres, &schema)
+                .build(&Filter::parse("task.deleted < false").unwrap())
+                .is_err()
+        );
     }
 }
