@@ -1,7 +1,7 @@
 use std::ops::Deref;
 
 use proc_macro2::{Ident, TokenStream};
-use quote::{quote, ToTokens};
+use quote::{ToTokens, quote};
 use syn::parse::{Parse, ParseStream};
 use syn::{
     ExprClosure, GenericArgument, Pat, PatType, Path, PathArguments, ReturnType, Token, Type,
@@ -41,26 +41,22 @@ pub fn expand(options: DerivedMap) -> syn::Result<TokenStream> {
 
     let mut is_request_result = false;
     let mut parse_return_type = None;
+
     if let ReturnType::Type(_, return_type) = &parse_item_closure.output {
         if let Type::Path(TypePath {
             path: Path { segments, .. },
             ..
         }) = &**return_type
-        {
-            if let Some(syn::PathSegment {
+            && let Some(syn::PathSegment {
                 ident,
                 arguments: PathArguments::AngleBracketed(args),
             }) = segments.first()
-            {
-                if let Some(GenericArgument::Type(ty)) = args.args.first() {
-                    if ident == "RequestResult" {
-                        is_request_result = true;
-                        parse_return_type = Some(ty.clone());
-                    }
-                }
-            }
-        }
-        if parse_return_type.is_none() {
+            && let Some(GenericArgument::Type(ty)) = args.args.first()
+            && ident == "RequestResult"
+        {
+            is_request_result = true;
+            parse_return_type = Some(ty.clone());
+        } else if parse_return_type.is_none() {
             parse_return_type = Some(return_type.deref().clone());
         }
     }
@@ -92,11 +88,10 @@ pub fn expand(options: DerivedMap) -> syn::Result<TokenStream> {
         ));
     };
 
-    let map_type = if let Some(map_type) = map_type {
-        map_type.into_token_stream()
-    } else {
-        quote! { ::std::collections::BTreeMap }
-    };
+    let map_type = map_type.map_or_else(
+        || quote! { ::std::collections::BTreeMap },
+        darling::ToTokens::into_token_stream,
+    );
 
     let parse_body = &parse_item_closure.body;
     let parse_expr = if is_request_result {
@@ -109,25 +104,25 @@ pub fn expand(options: DerivedMap) -> syn::Result<TokenStream> {
         }
     };
 
-    let write_fn = if let Some(write_item_closure) = write_item_closure.as_ref() {
-        let write_body = &write_item_closure.body;
-        quote! {
-            pub fn write<I>(values: I) -> Vec<#source_item_type>
-            where
-                I: ::std::iter::IntoIterator<Item = (#key_type, #value_type)>,
-            {
-                values
-                    .into_iter()
-                    .map(|item| {
-                        #[allow(unused_braces)]
-                        #write_body
-                    })
-                    .collect()
+    let write_fn = write_item_closure.as_ref().map_or_else(
+        || quote!(),
+        |write_item_closure| {
+            let write_body = &write_item_closure.body;
+            quote! {
+                pub fn write<I>(values: I) -> Vec<#source_item_type>
+                where
+                    I: ::std::iter::IntoIterator<Item = (#key_type, #value_type)>,
+                {
+                    values
+                        .into_iter()
+                        .map(|item| {
+                            #write_body
+                        })
+                        .collect()
+                }
             }
-        }
-    } else {
-        quote!()
-    };
+        },
+    );
 
     Ok(quote! {
         #vis mod #ident {
@@ -140,7 +135,6 @@ pub fn expand(options: DerivedMap) -> syn::Result<TokenStream> {
                 let mut m = #map_type::new();
                 for item in values.into_iter() {
                     let (k, v): (#key_type, #value_type) = {
-                        #[allow(unused_braces)]
                         #parse_expr
                     };
                     if m.insert(k, v).is_some() {
