@@ -1,6 +1,6 @@
 use bomboni_core::syn::type_is_phantom;
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote, ToTokens};
+use quote::{ToTokens, format_ident, quote};
 use std::collections::BTreeSet;
 
 use crate::parse::{
@@ -30,6 +30,13 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
         }
     }
 
+    // Parse field mask fields
+    for field in fields.iter().filter(|field| !field.options.skip) {
+        if let Some(field_mask) = &field.options.field_mask {
+            parse_fields.extend(expand_parse_field_mask(field, field_mask)?);
+        }
+    }
+
     // Parse regular fields
     for field in fields.iter().filter(|field| {
         !field.options.skip
@@ -37,6 +44,7 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
             && field.resource.is_none()
             && field.list_query.is_none()
             && field.search_query.is_none()
+            && field.options.field_mask.is_none()
             && !type_is_phantom(&field.ty)
     }) {
         parse_fields.extend(expand_parse_field(field, &field_clone_set)?);
@@ -48,7 +56,9 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
         .iter()
         .filter(|field| field.options.skip || type_is_phantom(&field.ty))
     {
-        let target_ident = field.ident.as_ref().unwrap();
+        let target_ident = field.ident.as_ref().ok_or_else(|| {
+            syn::Error::new(proc_macro2::Span::call_site(), "field missing ident")
+        })?;
         skipped_fields.extend(quote! {
             #target_ident: Default::default(),
         });
@@ -69,17 +79,22 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
         }
     }) {
         list_or_search = Some(field.list_query.is_some());
-        let target_ident = field.ident.as_ref().unwrap();
+        let target_ident = field.ident.as_ref().ok_or_else(|| {
+            syn::Error::new(proc_macro2::Span::call_site(), "field missing ident")
+        })?;
         let parse_query_impl = expand_parse_query(query, field.search_query.is_some());
-        query_token_type = if let Some(token_type) = get_query_field_token_type(&field.ty) {
-            quote! {
-                <PageToken = #token_type>
-            }
-        } else {
-            quote! {
-                <PageToken = FilterPageToken>
-            }
-        };
+        query_token_type = get_query_field_token_type(&field.ty).map_or_else(
+            || {
+                quote! {
+                    <PageToken = FilterPageToken>
+                }
+            },
+            |token_type| {
+                quote! {
+                    <PageToken = #token_type>
+                }
+            },
+        );
         quote! {
             Ok(Self {
                 #target_ident: {
@@ -104,11 +119,10 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
     let (impl_generics, type_generics, where_clause) = options.generics.split_for_impl();
 
     if let Some(request_options) = options.request.as_ref() {
-        let request_name = if let Some(name) = request_options.name.as_ref() {
-            quote! { #name }
-        } else {
-            quote! { #source::NAME }
-        };
+        let request_name = request_options
+            .name
+            .as_ref()
+            .map_or_else(|| quote! { #source::NAME }, |name| quote! { #name });
         parse_impl = quote! {
             (|| { #parse_impl })().map_err(|err: RequestError| err.wrap_request(#request_name))
         };
@@ -119,7 +133,6 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
             quote! {
                 #[automatically_derived]
                 impl #ident #type_generics #where_clause {
-                    #[allow(clippy::ignored_unit_patterns)]
                     pub fn parse_list_query<P: PageTokenBuilder #query_token_type >(
                         source: #source,
                         query_builder: &ListQueryBuilder<P>
@@ -132,7 +145,6 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
             quote! {
                 #[automatically_derived]
                 impl #ident #type_generics #where_clause {
-                    #[allow(clippy::ignored_unit_patterns)]
                     pub fn parse_search_query<P: PageTokenBuilder #query_token_type >(
                         source: #source,
                         query_builder: &SearchQueryBuilder<P>
@@ -145,7 +157,6 @@ pub fn expand(options: &ParseOptions, fields: &[ParseField]) -> syn::Result<Toke
             quote! {
                 #[automatically_derived]
                 impl #impl_generics RequestParse<#source> for #ident #type_generics #where_clause {
-                    #[allow(clippy::ignored_unit_patterns)]
                     fn parse(source: #source) -> RequestResult<Self> {
                         #parse_impl
                     }
@@ -159,7 +170,10 @@ fn expand_parse_field(
     field: &ParseField,
     field_clone_set: &BTreeSet<String>,
 ) -> syn::Result<TokenStream> {
-    let target_ident = field.ident.as_ref().unwrap();
+    let target_ident = field
+        .ident
+        .as_ref()
+        .ok_or_else(|| syn::Error::new(proc_macro2::Span::call_site(), "field missing ident"))?;
     let extract = get_field_extract(field)?;
 
     if let Some(ParseDerive {
@@ -213,7 +227,9 @@ fn expand_parse_field(
         ));
     }
 
-    let field_type_info = field.type_info.as_ref().unwrap();
+    let field_type_info = field.type_info.as_ref().ok_or_else(|| {
+        syn::Error::new(proc_macro2::Span::call_site(), "field missing type info")
+    })?;
     let (extract_impl, get_impl, field_path) = expand_field_extract(
         &extract,
         field_clone_set,
@@ -366,4 +382,133 @@ fn expand_parse_query(options: &ParseQuery, search: bool) -> TokenStream {
             let query = query_builder.build(page_size, page_token, filter, order_by)?;
         }
     }
+}
+
+fn expand_parse_field_mask(
+    field: &ParseField,
+    field_mask: &crate::parse::options::ParseFieldMask,
+) -> syn::Result<TokenStream> {
+    let target_ident = field
+        .ident
+        .as_ref()
+        .ok_or_else(|| syn::Error::new(proc_macro2::Span::call_site(), "field missing ident"))?;
+
+    // Extract container and field from source option
+    let (container_ident, field_path, field_optional) = field.options.source.as_ref().map_or_else(
+        || (None, target_ident.to_string(), false),
+        |source| {
+            // For source like "book?.display_name?", we want container="book", field="display_name", and track if field is optional
+            let parts: Vec<&str> = source.split('.').collect();
+            if parts.len() >= 2 {
+                let container_name = parts[0].trim_end_matches('?');
+                let field_name_raw = parts[1];
+                let field_optional = field_name_raw.ends_with('?');
+                let field_name = field_name_raw.trim_end_matches('?');
+                let container_ident = format_ident!("{}", container_name);
+                let field_path = field_name.to_string();
+                (Some(container_ident), field_path, field_optional)
+            } else {
+                (None, target_ident.to_string(), false)
+            }
+        },
+    );
+
+    // Determine container field for field mask
+    let _container_field = field_mask.field.as_ref().map_or_else(
+        || {
+            container_ident
+                .as_ref()
+                .map_or_else(|| format_ident!("book"), Clone::clone) // Default fallback
+        },
+        Clone::clone,
+    );
+
+    let mask_field = &field_mask.mask;
+
+    // Generate the field extraction logic
+    let extract_logic = container_ident.as_ref().map_or_else(
+        || {
+            quote! {
+                {
+                    let target = source.#target_ident;
+                    Some(target)
+                }
+            }
+        },
+        |container_ident| {
+            let field_ident = format_ident!("{}", field_path);
+            if field_optional {
+                quote! {
+                    {
+                        let target = source.#container_ident.clone();
+                        if let Some(target) = target {
+                            target.#field_ident
+                        } else {
+                            None
+                        }
+                    }
+                }
+            } else {
+                quote! {
+                    {
+                        let target = source.#container_ident.clone();
+                        if let Some(target) = target {
+                            Some(target.#field_ident)
+                        } else {
+                            None
+                        }
+                    }
+                }
+            }
+        },
+    );
+
+    // Generate field error path
+    let field_error_path = make_field_error_path(&field_path, None);
+
+    // Generate the complete parsing logic with field mask check
+    field.options.convert.as_ref().map_or_else(|| Ok(quote! {
+        #target_ident: {
+            // Check if field mask allows this field
+            if let Some(mask) = source.#mask_field.as_ref() {
+                if mask.paths.contains(&#field_path.to_string()) {
+                    #extract_logic
+                } else {
+                    None
+                }
+            } else {
+                // No field mask provided, don't parse this field
+                None
+            }
+        },
+    }), |convert| {
+        let convert_impl = convert
+            .parse
+            .as_ref()
+            .map(ToTokens::to_token_stream)
+            .or_else(|| convert.module.as_ref().map(|module| quote!(#module::parse)))
+            .unwrap();
+
+        Ok(quote! {
+            #target_ident: {
+                // Check if field mask allows this field
+                if let Some(mask) = source.#mask_field.as_ref() {
+                    if mask.paths.contains(&#field_path.to_string()) {
+                        let target = #extract_logic;
+                        let target = if let Some(target) = target {
+                            Some(#convert_impl(target).map_err(|err: RequestError| err.wrap_path(#field_error_path))?)
+                        } else {
+                            None
+                        };
+                        target
+                    } else {
+                        None
+                    }
+                } else {
+                    // No field mask provided, don't parse this field
+                    None
+                }
+            },
+        })
+    })
 }
