@@ -1,7 +1,7 @@
 use std::collections::BTreeSet;
 
 use proc_macro2::TokenStream;
-use quote::{quote, ToTokens};
+use quote::{ToTokens, quote};
 use syn::Path;
 
 use crate::parse::{
@@ -12,11 +12,10 @@ use crate::parse::{
 };
 
 pub fn expand(options: &ParseOptions, variants: &[ParseVariant]) -> syn::Result<TokenStream> {
-    if let Some(tagged_union) = options.tagged_union.as_ref() {
-        expand_tagged_union(options, variants, tagged_union)
-    } else {
-        expand_parse(options, variants)
-    }
+    options.tagged_union.as_ref().map_or_else(
+        || expand_parse(options, variants),
+        |tagged_union| expand_tagged_union(options, variants, tagged_union),
+    )
 }
 
 fn expand_parse(options: &ParseOptions, variants: &[ParseVariant]) -> syn::Result<TokenStream> {
@@ -64,7 +63,6 @@ fn expand_tagged_union(
     Ok(quote! {
         #[automatically_derived]
         impl #impl_generics RequestParse<#source> for #ident #type_generics #where_clause {
-            #[allow(ignored_unit_patterns)]
             fn parse(source: #source) -> RequestResult<Self> {
                 let source = source.#field_ident
                     .ok_or_else(|| RequestError::field(#field_literal, CommonError::RequiredFieldMissing))?;
