@@ -23,19 +23,18 @@ pub fn get_variant_extract(variant: &ParseVariant) -> syn::Result<FieldExtract> 
         steps.extend(extract.steps);
     }
 
-    if let Some(field_type_info) = variant.type_info.as_ref() {
-        if variant.options.extract.is_none()
-            && variant.options.source.is_none()
-            && variant.options.derive.is_none()
-            && !variant.options.oneof
-            && !variant.options.enumeration
-        {
-            if field_type_info.container_ident.as_deref() == Some("Option") {
-                steps.push(FieldExtractStep::Unwrap);
-            }
-            if field_type_info.container_ident.as_deref() == Some("Box") {
-                steps.push(FieldExtractStep::Unbox);
-            }
+    if let Some(field_type_info) = variant.type_info.as_ref()
+        && variant.options.extract.is_none()
+        && variant.options.source.is_none()
+        && variant.options.derive.is_none()
+        && !variant.options.oneof
+        && !variant.options.enumeration
+    {
+        if field_type_info.container_ident.as_deref() == Some("Option") {
+            steps.push(FieldExtractStep::Unwrap);
+        }
+        if field_type_info.container_ident.as_deref() == Some("Box") {
+            steps.push(FieldExtractStep::Unbox);
         }
     }
 
