@@ -6,32 +6,54 @@ use std::collections::BTreeMap;
 
 use super::utility::get_param;
 
+/// Name of upper case helper.
 pub const UPPER_CASE_HELPER_NAME: &str = "upperCase";
+/// Name of lower case helper.
 pub const LOWER_CASE_HELPER_NAME: &str = "lowerCase";
+/// Name of title case helper.
 pub const TITLE_CASE_HELPER_NAME: &str = "titleCase";
+/// Name of toggle case helper.
 pub const TOGGLE_CASE_HELPER_NAME: &str = "toggleCase";
+/// Name of alternating case helper.
 pub const ALTERNATING_CASE_HELPER_NAME: &str = "alternatingCase";
+/// Name of camel case helper.
 pub const CAMEL_CASE_HELPER_NAME: &str = "camelCase";
+/// Name of pascal case helper.
 pub const PASCAL_CASE_HELPER_NAME: &str = "pascalCase";
+/// Name of upper camel case helper.
 pub const UPPER_CAMEL_CASE_HELPER_NAME: &str = "upperCamelCase";
+/// Name of snake case helper.
 pub const SNAKE_CASE_HELPER_NAME: &str = "snakeCase";
+/// Name of upper snake case helper.
 pub const UPPER_SNAKE_CASE_HELPER_NAME: &str = "upperSnakeCase";
+/// Name of screaming snake case helper.
 pub const SCREAMING_SNAKE_CASE_HELPER_NAME: &str = "screamingSnakeCase";
+/// Name of kebab case helper.
 pub const KEBAB_CASE_HELPER_NAME: &str = "kebabCase";
+/// Name of cobol case helper.
 pub const COBOL_CASE_HELPER_NAME: &str = "cobolCase";
+/// Name of train case helper.
 pub const TRAIN_CASE_HELPER_NAME: &str = "trainCase";
+/// Name of flat case helper.
 pub const FLAT_CASE_HELPER_NAME: &str = "flatCase";
+/// Name of upper flat case helper.
 pub const UPPER_FLAT_CASE_HELPER_NAME: &str = "upperFlatCase";
 
+/// Name of to string helper.
 pub const TO_STRING_HELPER_NAME: &str = "toString";
+/// Name of to JSON helper.
 pub const TO_JSON_HELPER_NAME: &str = "toJson";
+/// Name of concat helper.
 pub const CONCAT_HELPER_NAME: &str = "concat";
+/// Name of to integer string helper.
 pub const TO_INTEGER_STRING_HELPER_NAME: &str = "toIntegerString";
 
+/// Registers all string helpers with Handlebars registry.
 pub fn register_string_helpers(handlebars_registry: &mut Handlebars) {
     register_string_helpers_with_name_map(handlebars_registry, BTreeMap::default());
 }
 
+/// Registers string helpers with custom name mapping.
 pub fn register_string_helpers_with_name_map(
     handlebars_registry: &mut Handlebars,
     name_map: BTreeMap<String, String>,
@@ -87,11 +109,9 @@ fn convert_case_helper(
     out: &mut dyn Output,
 ) -> HelperResult {
     let value = get_param_value(h, 0, "value")?;
-    let value = if let Some(s) = value.as_str() {
-        s.to_string()
-    } else {
-        value.to_string()
-    };
+    let value = value
+        .as_str()
+        .map_or_else(|| value.to_string(), ToString::to_string);
     out.write(&match h.name() {
         UPPER_CASE_HELPER_NAME => value.to_case(Case::Upper),
         LOWER_CASE_HELPER_NAME => value.to_case(Case::Lower),
@@ -103,7 +123,7 @@ fn convert_case_helper(
         UPPER_CAMEL_CASE_HELPER_NAME => value.to_case(Case::UpperCamel),
         SNAKE_CASE_HELPER_NAME => value.to_case(Case::Snake),
         UPPER_SNAKE_CASE_HELPER_NAME => value.to_case(Case::UpperSnake),
-        SCREAMING_SNAKE_CASE_HELPER_NAME => value.to_case(Case::ScreamingSnake),
+        SCREAMING_SNAKE_CASE_HELPER_NAME => value.to_case(Case::Constant),
         KEBAB_CASE_HELPER_NAME => value.to_case(Case::Kebab),
         COBOL_CASE_HELPER_NAME => value.to_case(Case::Cobol),
         TRAIN_CASE_HELPER_NAME => value.to_case(Case::Train),
