@@ -1,10 +1,10 @@
-//! # Wasm Core
-//!
-//! Based on [1] and [2].
-//!
-//! [1]: https://github.com/madonoharu/tsify
-//! [2]: https://github.com/Aleph-Alpha/ts-rs
+#![doc = include_str!("../README.md")]
 
+/// Options for configuring Wasm bindings.
 pub mod options;
+
+/// TypeScript declaration generation.
 pub mod ts_decl;
+
+/// TypeScript type representations.
 pub mod ts_type;
